master: save-commit before file transfer on new project to fix start page routing bug

+++ app/core/component/course-list/courses-list.controller.js
@@ -30,7 +30,10 @@ angular
         };
 
         $ctrl.deleteCourse = function(course) {
-            courseService.deleteCourse(course);
+            // eslint-disable-next-line no-alert
+            if (confirm('Do you really want to delete this course?')) {
+                courseService.deleteCourse(course);
+            }
         };
 
         $ctrl.isTopRated = function(course) {
